+++ nbi/engine/src/org/netbeans/installer/utils/system/launchers/impl/ExeLauncher.java
@@ -325,7 +325,7 @@ public class ExeLauncher extends CommonLauncher {
     private void addData(FileOutputStream fos, String [] strings, boolean isUnicode) throws IOException {
         
         if(strings!=null) {
-            addNumber(fos, Integer.valueOf(strings.length).longValue()); // number of array elements
+            addNumber(fos, strings.length); // number of array elements
             for(String s: strings) {
                 addData(fos, s, isUnicode);
             }
