+++ php/php.editor/test/unit/src/org/netbeans/modules/php/editor/indent/PHPFormatterTestBase.java
@@ -96,7 +96,7 @@ public abstract class PHPFormatterTestBase extends PHPTestBase {
             String option = entry.getKey();
             Object value = entry.getValue();
             if (value instanceof Integer) {
-                prefs.putInt(option, ((Integer) value).intValue());
+                prefs.putInt(option, (Integer) value);
             } else if (value instanceof String) {
                 prefs.put(option, (String) value);
             } else if (value instanceof Boolean) {
@@ -180,7 +180,7 @@ public abstract class PHPFormatterTestBase extends PHPTestBase {
             String option = entry.getKey();
             Object value = entry.getValue();
             if (value instanceof Integer) {
-                prefs.putInt(option, ((Integer) value).intValue());
+                prefs.putInt(option, (Integer) value);
             } else if (value instanceof String) {
                 prefs.put(option, (String) value);
             } else if (value instanceof Boolean) {
