+++ platform/o.n.swing.tabcontrol/src/org/netbeans/swing/tabcontrol/plaf/DefaultTabbedContainerUI.java
@@ -1211,7 +1211,7 @@ public class DefaultTabbedContainerUI extends TabbedContainerUI {
                             //Get the index into the old array of a deleted tab
                             Integer idx = (Integer) i.next();
                             //Find the TabData object for it
-                            TabData del = old[idx.intValue()];
+                            TabData del = old[idx];
                             //Make sure its component is not one we'll be adding
                             if (!components.contains(toComp(del))) {
                                 //remove it
@@ -1230,7 +1230,7 @@ public class DefaultTabbedContainerUI extends TabbedContainerUI {
                                 //Get the index into the new array of the added tab
                                 Integer idx = (Integer) i.next();
                                 //Find the TabData object that was added
-                                TabData add = nue[idx.intValue()];
+                                TabData add = nue[idx];
                                 //Make sure it's not already showing so we don't do
                                 //extra work
                                 if (!contentDisplayer.isAncestorOf(
